+++ sentry/src/transports/reqwest.rs
@@ -1,9 +1,11 @@
 use std::time::Duration;
 
 use reqwest::{header as ReqwestHeaders, Client as ReqwestClient, Proxy, StatusCode};
+use sentry_core::TransportOptions;
 
 use super::{
-    tokio_thread::TransportThread, HTTP_PAYLOAD_TOO_LARGE, HTTP_PAYLOAD_TOO_LARGE_MESSAGE,
+    tokio_thread::{TransportThread, TransportThreadOptions},
+    RateLimiter, HTTP_PAYLOAD_TOO_LARGE, HTTP_PAYLOAD_TOO_LARGE_MESSAGE,
 };
 
 use crate::{sentry_debug, ClientOptions, Envelope, Transport};
@@ -18,24 +20,76 @@ pub struct ReqwestHttpTransport {
     thread: TransportThread,
 }
 
+/// Options for constructing a [`ReqwestHttpTransport`].
+///
+/// Currently, this is primarily a wrapper around a [`TransportOptions`], and must be created with
+/// the `From<TransportOptions>` implementation. Optionally, a [`reqwest::Client`] for the
+/// transport may be provided with [`Self::with_client`].
+#[derive(Debug)]
+#[must_use]
+pub struct ReqwestHttpTransportOptions {
+    general_options: TransportOptions,
+    client: Option<ReqwestClient>,
+}
+
 impl ReqwestHttpTransport {
-    /// Creates a new Transport.
+    /// Backwards-compatible method for creating a [`ReqwestHttpTransport`].
+    ///
+    /// Please use [`ReqwestHttpTransportOptions::build`] instead.
+    ///
+    /// ### Panics
+    ///
+    /// Panics if called with `options` that lack a DSN.
+    #[inline]
+    #[deprecated = "use `ReqwestHttpTransportOptions::build` instead"]
     pub fn new(options: &ClientOptions) -> Self {
-        Self::new_internal(options, None)
+        let general_options = TransportOptions::try_from_client_options(options)
+            .expect("this method should only be called when options has a DSN");
+
+        ReqwestHttpTransportOptions::from(general_options).build()
     }
 
-    /// Creates a new Transport that uses the specified [`ReqwestClient`].
+    /// Backwards-compatible method for creating a [`ReqwestHttpTransport`] that uses the specified
+    /// [`ReqwestClient`].
+    ///
+    /// Please use [`ReqwestHttpTransportOptions::build`] instead.
+    ///
+    /// ### Panics
+    ///
+    /// Panics if called with `options` that lack a DSN.
+    #[inline]
+    #[deprecated = "use `ReqwestHttpTransportOptions::build` instead"]
     pub fn with_client(options: &ClientOptions, client: ReqwestClient) -> Self {
-        Self::new_internal(options, Some(client))
+        let general_options = TransportOptions::try_from_client_options(options)
+            .expect("this method should only be called when options has a DSN");
+
+        ReqwestHttpTransportOptions::from(general_options)
+            .with_client(client)
+            .build()
     }
 
-    fn new_internal(options: &ClientOptions, client: Option<ReqwestClient>) -> Self {
+    /// Creates a new [`ReqwestHttpTransport`] with the given `options`.
+    #[inline]
+    pub(super) fn with_options(options: ReqwestHttpTransportOptions) -> Self {
+        let ReqwestHttpTransportOptions {
+            general_options:
+                TransportOptions {
+                    dsn,
+                    user_agent,
+                    http_proxy,
+                    https_proxy,
+                    accept_invalid_certs,
+                    ..
+                },
+            client,
+        } = options;
+
         let client = client.unwrap_or_else(|| {
             let mut builder = reqwest::Client::builder();
-            if options.accept_invalid_certs {
+            if accept_invalid_certs {
                 builder = builder.danger_accept_invalid_certs(true);
             }
-            if let Some(url) = options.http_proxy.as_ref() {
+            if let Some(url) = http_proxy.as_ref() {
                 match Proxy::http(url.as_ref()) {
                     Ok(proxy) => {
                         builder = builder.proxy(proxy);
@@ -45,7 +99,7 @@ impl ReqwestHttpTransport {
                     }
                 }
             };
-            if let Some(url) = options.https_proxy.as_ref() {
+            if let Some(url) = https_proxy.as_ref() {
                 match Proxy::https(url.as_ref()) {
                     Ok(proxy) => {
                         builder = builder.proxy(proxy);
@@ -59,12 +113,11 @@ impl ReqwestHttpTransport {
                 .build()
                 .expect("Failed to build `reqwest` client as a TLS backend is not available. Enable either the `native-tls` or the `rustls` feature of the `sentry` crate.")
         });
-        let dsn = options.dsn.as_ref().unwrap();
-        let user_agent = options.user_agent.clone();
+
         let auth = dsn.to_auth(Some(&user_agent)).to_string();
         let url = dsn.envelope_api_url().to_string();
 
-        let thread = TransportThread::new(move |envelope, mut rl| {
+        let send_fn = move |envelope: Envelope, mut rl: RateLimiter| {
             let mut body = Vec::new();
             envelope.to_writer(&mut body).unwrap();
             let request = client.post(&url).header("X-Sentry-Auth", &auth).body(body);
@@ -110,7 +163,9 @@ impl ReqwestHttpTransport {
                 }
                 rl
             }
-        });
+        };
+
+        let thread = TransportThreadOptions::new(send_fn).spawn_thread();
         Self { thread }
     }
 }
@@ -125,5 +180,30 @@ impl Transport for ReqwestHttpTransport {
 
     fn shutdown(&self, timeout: Duration) -> bool {
         self.flush(timeout)
+    }
+}
+
+impl From<TransportOptions> for ReqwestHttpTransportOptions {
+    #[inline]
+    fn from(value: TransportOptions) -> Self {
+        Self {
+            general_options: value,
+            client: None,
+        }
+    }
+}
+
+impl ReqwestHttpTransportOptions {
+    /// Specify the [`reqwest::Client`] for the [`ReqwestHttpTransport`].
+    #[inline]
+    pub fn with_client(self, client: ReqwestClient) -> Self {
+        let client = Some(client);
+        Self { client, ..self }
+    }
+
+    /// Create a [`ReqwestHttpTransport`] using these options.
+    #[inline]
+    pub fn build(self) -> ReqwestHttpTransport {
+        ReqwestHttpTransport::with_options(self)
     }
 }