+++ sentry/src/transports/ureq.rs
@@ -1,5 +1,6 @@
 use std::time::Duration;
 
+use sentry_core::TransportOptions;
 use ureq::http::Response;
 #[cfg(any(
     feature = "rustls",
@@ -9,7 +10,10 @@ use ureq::http::Response;
 use ureq::tls::{TlsConfig, TlsProvider};
 use ureq::{Agent, Proxy};
 
-use super::{thread::TransportThread, HTTP_PAYLOAD_TOO_LARGE, HTTP_PAYLOAD_TOO_LARGE_MESSAGE};
+use super::{
+    thread::{TransportThread, TransportThreadOptions},
+    RateLimiter, HTTP_PAYLOAD_TOO_LARGE, HTTP_PAYLOAD_TOO_LARGE_MESSAGE,
+};
 
 use crate::{sentry_debug, types::Scheme, ClientOptions, Envelope, Transport};
 
@@ -21,19 +25,74 @@ pub struct UreqHttpTransport {
     thread: TransportThread,
 }
 
+/// Options for constructing a [`UreqHttpTransport`].
+///
+/// Currently, this is primarily a wrapper around a [`TransportOptions`], and must be created with
+/// the `From<TransportOptions>` implementation. Optionally, a [`ureq::Agent`] for the transport may
+/// be provided with [`Self::with_agent`].
+#[derive(Debug)]
+#[must_use]
+pub struct UreqHttpTransportOptions {
+    general_options: TransportOptions,
+    agent: Option<Agent>,
+}
+
 impl UreqHttpTransport {
-    /// Creates a new Transport.
+    /// Backwards-compatible method for creating a [`UreqHttpTransport`].
+    ///
+    /// Please use [`UreqHttpTransportOptions::build`] instead.
+    ///
+    /// ### Panics
+    ///
+    /// Panics if called with `options` that lack a DSN.
+    #[inline]
+    #[deprecated = "use `UreqHttpTransportOptions::build` instead"]
     pub fn new(options: &ClientOptions) -> Self {
-        Self::new_internal(options, None)
+        let general_options = TransportOptions::try_from_client_options(options)
+            .expect("this method should only be called when options has a DSN");
+
+        UreqHttpTransportOptions::from(general_options).build()
     }
 
-    /// Creates a new Transport that uses the specified [`ureq::Agent`].
+    /// Backwards-compatible method for creating a [`UreqHttpTransport`] that uses the specified
+    /// [`ureq::Agent`].
+    ///
+    /// Please use [`UreqHttpTransportOptions::build`] instead.
+    ///
+    /// ### Panics
+    ///
+    /// Panics if called with `options` that lack a DSN.
+    #[inline]
+    #[deprecated = "use `UreqHttpTransportOptions::build` instead"]
     pub fn with_agent(options: &ClientOptions, agent: Agent) -> Self {
-        Self::new_internal(options, Some(agent))
+        let general_options = TransportOptions::try_from_client_options(options)
+            .expect("this method should only be called when options has a DSN");
+
+        UreqHttpTransportOptions::from(general_options)
+            .with_agent(agent)
+            .build()
     }
 
-    fn new_internal(options: &ClientOptions, agent: Option<Agent>) -> Self {
-        let dsn = options.dsn.as_ref().unwrap();
+    /// Creates a new [`UreqHttpTransport`] with the given `options`.
+    #[inline]
+    pub(super) fn with_options(options: UreqHttpTransportOptions) -> Self {
+        let UreqHttpTransportOptions {
+            general_options:
+                TransportOptions {
+                    dsn,
+                    user_agent,
+                    http_proxy,
+                    https_proxy,
+                    #[cfg(any(
+                        feature = "native-tls",
+                        feature = "rustls",
+                        feature = "rustls-no-provider"
+                    ))]
+                    accept_invalid_certs,
+                    ..
+                },
+            agent,
+        } = options;
         let scheme = dsn.scheme();
         let agent = agent.unwrap_or_else(|| {
             let mut builder = Agent::config_builder();
@@ -43,7 +102,7 @@ impl UreqHttpTransport {
                 builder = builder.tls_config(
                     TlsConfig::builder()
                         .provider(TlsProvider::NativeTls)
-                        .disable_verification(options.accept_invalid_certs)
+                        .disable_verification(accept_invalid_certs)
                         .build(),
                 );
             }
@@ -52,15 +111,15 @@ impl UreqHttpTransport {
                 builder = builder.tls_config(
                     TlsConfig::builder()
                         .provider(TlsProvider::Rustls)
-                        .disable_verification(options.accept_invalid_certs)
+                        .disable_verification(accept_invalid_certs)
                         .build(),
                 );
             }
 
             let mut maybe_proxy = None;
 
-            match (scheme, &options.http_proxy, &options.https_proxy) {
-                (Scheme::Https, _, Some(proxy)) => match Proxy::new(proxy) {
+            match (scheme, &http_proxy, &https_proxy) {
+                (Scheme::Https, _, Some(proxy)) => match Proxy::new(proxy.as_ref()) {
                     Ok(proxy) => {
                         maybe_proxy = Some(proxy);
                     }
@@ -68,7 +127,7 @@ impl UreqHttpTransport {
                         sentry_debug!("invalid proxy: {:?}", err);
                     }
                 },
-                (_, Some(proxy), _) => match Proxy::new(proxy) {
+                (_, Some(proxy), _) => match Proxy::new(proxy.as_ref()) {
                     Ok(proxy) => {
                         maybe_proxy = Some(proxy);
                     }
@@ -83,11 +142,10 @@ impl UreqHttpTransport {
 
             builder.build().new_agent()
         });
-        let user_agent = options.user_agent.clone();
         let auth = dsn.to_auth(Some(&user_agent)).to_string();
         let url = dsn.envelope_api_url().to_string();
 
-        let thread = TransportThread::new(move |envelope, rl| {
+        let send_fn = move |envelope: Envelope, rl: &mut RateLimiter| {
             let mut body = Vec::new();
             envelope.to_writer(&mut body).unwrap();
             let request = agent
@@ -128,7 +186,9 @@ impl UreqHttpTransport {
                     sentry_debug!("Failed to send envelope: {}", err);
                 }
             }
-        });
+        };
+
+        let thread = TransportThreadOptions::new(send_fn).spawn_thread();
         Self { thread }
     }
 }
@@ -143,5 +203,30 @@ impl Transport for UreqHttpTransport {
 
     fn shutdown(&self, timeout: Duration) -> bool {
         self.flush(timeout)
+    }
+}
+
+impl From<TransportOptions> for UreqHttpTransportOptions {
+    #[inline]
+    fn from(value: TransportOptions) -> Self {
+        Self {
+            general_options: value,
+            agent: None,
+        }
+    }
+}
+
+impl UreqHttpTransportOptions {
+    /// Specify the [`ureq::Agent`] for the [`UreqHttpTransport`].
+    #[inline]
+    pub fn with_agent(self, agent: Agent) -> Self {
+        let agent = Some(agent);
+        Self { agent, ..self }
+    }
+
+    /// Create a [`UreqHttpTransport`] using these options.
+    #[inline]
+    pub fn build(self) -> UreqHttpTransport {
+        UreqHttpTransport::with_options(self)
     }
 }